ecommerce/requirements.txt
absl-py==1.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.3.1
altair==4.2.2
annotated-types==0.7.0
ansicon==1.89.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
ariadne==0.26.1
arrow==1.3.0
arviz==0.15.1
asgiref==3.7.2
astropy==5.3.4
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
atpublic==4.1.0
attrs==23.2.0
audioread==3.0.1
autograd==1.6.2
bleach==6.2.0
blessed==1.20.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
chargily-epay-Django==0.0.2
chargily-epay-python==0.0.2
chargily_pay==2.1.3
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.9
Django==4.2.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-picklefield==3.3
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
entrypoints==0.4
executing==2.0.1
filelock==3.12.2
Flask==3.0.3
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.40.0
frozenlist==1.4.1
future==1.0.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
graphql-core==3.2.5
greenlet==3.1.1
grpcio==1.64.1
h5netcdf==1.3.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jinxed==1.3.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaggle==1.6.17
kagglehub==0.3.10
keras==2.15.0
kiwisolver==1.4.4
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.5
mysql-connector-python==8.0.33
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==23.1
pandas==2.0.3
parso==0.8.4
Pillow==10.0.0
pipenv==2024.4.1
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
PuLP==3.1.1
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyerfa==2.0.1.4
pygame==2.5.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
qudida==0.0.4
redis==3.5.3
referencing==0.35.1
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.46.1
stripe==12.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.6.18
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.20
virtualenv==20.29.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.14.1
xarray==2024.6.0
xarray-einstats==0.7.0
yarl==1.9.4
